Remove `ContextProviderType`, replace with `ContextProvider`
